+++ backend/canisters/group/impl/src/lifecycle/post_upgrade.rs
@@ -1,13 +1,16 @@
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
+use crate::timer_job_types::MigrateMembersToStableMemoryJob;
 use crate::{read_state, Data};
 use canister_logger::LogEntry;
+use canister_timer_jobs::Job;
 use canister_tracing_macros::trace;
 use group_canister::post_upgrade::Args;
 use ic_cdk::post_upgrade;
 use instruction_counts_log::InstructionCountFunctionId;
 use stable_memory::get_reader;
 use tracing::info;
+use types::MultiUserChat;
 
 #[post_upgrade]
 #[trace]
@@ -21,6 +24,7 @@ fn post_upgrade(args: Args) {
         msgpack::deserialize(reader).unwrap();
 
     data.chat.members.set_member_default_timestamps();
+    data.chat.members.set_chat(MultiUserChat::Group(ic_cdk::id().into()));
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
@@ -35,4 +39,6 @@ fn post_upgrade(args: Args) {
             .data
             .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now)
     });
+
+    MigrateMembersToStableMemoryJob.execute();
 }