+++ backend/libraries/chat_events/src/chat_events.rs
@@ -3,7 +3,6 @@ use crate::expiring_events::ExpiringEvents;
 use crate::last_updated_timestamps::LastUpdatedTimestamps;
 use crate::metrics::{ChatMetricsInternal, MetricKey};
 use crate::search_index::SearchIndex;
-use crate::stable_memory::key::KeyPrefix;
 use crate::*;
 use constants::{HOUR_IN_MS, OPENCHAT_BOT_USER_ID};
 use event_store_producer::{EventBuilder, EventStoreClient, Runtime};
@@ -82,10 +81,6 @@ impl ChatEvents {
 
     pub fn import_events(chat: Chat, events: Vec<(EventContext, ByteBuf)>) {
         stable_memory::write_events_as_bytes(chat, events);
-    }
-
-    pub fn garbage_collect_stable_memory(prefix: KeyPrefix) -> Result<u32, u32> {
-        stable_memory::garbage_collect(prefix)
     }
 
     pub fn set_stable_memory_key_prefixes(&mut self) {